Make use of extract state modules
