+++ domains/android/compose-core/build.gradle
@@ -5,5 +5,4 @@ dependencies {
     implementation project(":features:navigator")
     implementation project(":design-library")
     api project(":domains:android:core")
-    api project(":domains:state")
 }
