+++ domains/android/compose-core/src/main/kotlin/app/dapk/st/core/ActivityExtensions.kt
@@ -21,53 +21,3 @@ inline fun <reified VM : ViewModel> ComponentActivity.viewModel(
     }
     return ViewModelLazy(VM::class, { viewModelStore }, { factoryPromise })
 }
-
-
-inline fun <reified S, E> ComponentActivity.state(
-    noinline factory: () -> State<S, E>
-): Lazy<State<S, E>> {
-    val factoryPromise = object : Factory {
-        @Suppress("UNCHECKED_CAST")
-        override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return when(modelClass) {
-                StateViewModel::class.java -> factory() as T
-                else -> throw Error()
-            }
-        }
-    }
-    return KeyedViewModelLazy(
-        key = S::class.java.canonicalName!!,
-        StateViewModel::class,
-        { viewModelStore },
-        { factoryPromise }
-    ) as Lazy<State<S, E>>
-}
-
-class KeyedViewModelLazy<VM : ViewModel> @JvmOverloads constructor(
-    private val key: String,
-    private val viewModelClass: KClass<VM>,
-    private val storeProducer: () -> ViewModelStore,
-    private val factoryProducer: () -> ViewModelProvider.Factory,
-) : Lazy<VM> {
-    private var cached: VM? = null
-
-    override val value: VM
-        get() {
-            val viewModel = cached
-            return if (viewModel == null) {
-                val factory = factoryProducer()
-                val store = storeProducer()
-                ViewModelProvider(
-                    store,
-                    factory,
-                    CreationExtras.Empty
-                ).get(key, viewModelClass.java).also {
-                    cached = it
-                }
-            } else {
-                viewModel
-            }
-        }
-
-    override fun isInitialized(): Boolean = cached != null
-}
