+++ features/directory/build.gradle
@@ -3,7 +3,7 @@ applyAndroidComposeLibraryModule(project)
 dependencies {
     implementation project(":chat-engine")
     implementation project(":domains:android:compose-core")
-    implementation project(":domains:state")
+    implementation 'screen-state:screen-android'
     implementation project(":features:messenger")
     implementation project(":core")
     implementation project(":design-library")
@@ -11,9 +11,9 @@ dependencies {
 
     kotlinTest(it)
 
+    testImplementation 'screen-state:state-test'
     androidImportFixturesWorkaround(project, project(":matrix:common"))
     androidImportFixturesWorkaround(project, project(":core"))
-    androidImportFixturesWorkaround(project, project(":domains:state"))
     androidImportFixturesWorkaround(project, project(":domains:store"))
     androidImportFixturesWorkaround(project, project(":domains:android:stub"))
     androidImportFixturesWorkaround(project, project(":chat-engine"))
