+++ features/directory/src/main/kotlin/app/dapk/st/directory/DirectoryModule.kt
@@ -2,7 +2,7 @@ package app.dapk.st.directory
 
 import android.content.Context
 import app.dapk.st.core.ProvidableModule
-import app.dapk.st.core.createStateViewModel
+import app.dapk.st.state.createStateViewModel
 import app.dapk.st.core.JobBag
 import app.dapk.st.directory.state.DirectoryState
 import app.dapk.st.directory.state.directoryReducer
