+++ features/home/build.gradle
@@ -9,7 +9,7 @@ dependencies {
     implementation project(":domains:android:compose-core")
     implementation project(":domains:android:viewmodel")
     implementation project(':domains:store')
-    implementation project(':domains:state')
+    implementation 'screen-state:screen-android'
     implementation project(":core")
     implementation project(":design-library")
     implementation Dependencies.mavenCentral.coil
