+++ features/messenger/src/main/kotlin/app/dapk/st/messenger/MessengerModule.kt
@@ -5,7 +5,7 @@ import android.content.Context
 import app.dapk.st.core.DeviceMeta
 import app.dapk.st.core.JobBag
 import app.dapk.st.core.ProvidableModule
-import app.dapk.st.core.createStateViewModel
+import app.dapk.st.state.createStateViewModel
 import app.dapk.st.domain.application.message.MessageOptionsStore
 import app.dapk.st.engine.ChatEngine
 import app.dapk.st.matrix.common.RoomId
