+++ features/messenger/src/main/kotlin/app/dapk/st/messenger/gallery/state/ImageGalleryState.kt
@@ -1,12 +1,12 @@
 package app.dapk.st.messenger.gallery.state
 
 import app.dapk.st.core.Lce
-import app.dapk.st.core.State
-import app.dapk.st.design.components.Route
+import app.dapk.st.state.State
 import app.dapk.st.messenger.gallery.Folder
 import app.dapk.st.messenger.gallery.Media
-import app.dapk.st.core.page.PageContainer
 import app.dapk.state.Combined2
+import app.dapk.state.Route
+import app.dapk.state.page.PageContainer
 
 typealias ImageGalleryState = State<Combined2<PageContainer<ImageGalleryPage>, Unit>, Unit>
