+++ features/profile/build.gradle
@@ -4,16 +4,16 @@ dependencies {
     implementation project(":chat-engine")
     implementation project(":features:settings")
     implementation project(':domains:store')
-    implementation project(':domains:state')
+    implementation 'screen-state:screen-android'
     implementation project(":domains:android:compose-core")
     implementation project(":design-library")
     implementation project(":core")
 
     kotlinTest(it)
 
+    testImplementation 'screen-state:state-test'
     androidImportFixturesWorkaround(project, project(":matrix:common"))
     androidImportFixturesWorkaround(project, project(":core"))
-    androidImportFixturesWorkaround(project, project(":domains:state"))
     androidImportFixturesWorkaround(project, project(":domains:store"))
     androidImportFixturesWorkaround(project, project(":domains:android:stub"))
     androidImportFixturesWorkaround(project, project(":chat-engine"))
