+++ features/profile/src/main/kotlin/app/dapk/st/profile/state/ProfileReducer.kt
@@ -3,14 +3,14 @@ package app.dapk.st.profile.state
 import app.dapk.st.core.JobBag
 import app.dapk.st.core.Lce
 import app.dapk.st.core.extensions.ErrorTracker
-import app.dapk.st.core.page.PageAction
-import app.dapk.st.core.page.PageStateChange
-import app.dapk.st.core.page.createPageReducer
-import app.dapk.st.core.page.withPageContext
-import app.dapk.st.design.components.SpiderPage
 import app.dapk.st.engine.ChatEngine
+import app.dapk.state.SpiderPage
 import app.dapk.state.async
 import app.dapk.state.createReducer
+import app.dapk.state.page.PageAction
+import app.dapk.state.page.PageStateChange
+import app.dapk.state.page.createPageReducer
+import app.dapk.state.page.withPageContext
 import app.dapk.state.sideEffect
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
