+++ features/settings/build.gradle
@@ -7,15 +7,16 @@ dependencies {
     implementation project(':domains:android:push')
     implementation project(":domains:android:compose-core")
     implementation project(":domains:android:viewmodel")
+    implementation 'screen-state:screen-android'
     implementation project(":design-library")
     implementation project(":core")
 
     kotlinTest(it)
 
+    testImplementation 'screen-state:state-test'
     androidImportFixturesWorkaround(project, project(":matrix:common"))
     androidImportFixturesWorkaround(project, project(":core"))
     androidImportFixturesWorkaround(project, project(":domains:store"))
-    androidImportFixturesWorkaround(project, project(":domains:state"))
     androidImportFixturesWorkaround(project, project(":domains:android:viewmodel"))
     androidImportFixturesWorkaround(project, project(":domains:android:stub"))
     androidImportFixturesWorkaround(project, project(":chat-engine"))
